requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.5
gpib-ctypes==0.3.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pillow==11.3.0
pyusb==1.2.1
pywin32==306
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-vxi11==0.9
PyVISA==1.15.0
PyVISA-py==0.8.0
six==1.17.0
typing_extensions==4.14.1
